portfolio/02-document-ai/requirements.txt
easyocr>=1.7.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pillow>=10.0.0
opencv-python>=4.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
